[RUBY-4194] extracted some of the logic from InitialComplianceChargeRowPresenter and AdditionalComplianceChargeRowPresenter into a parent class ComplianceChargeRowPresenter and fixed an issue with balance not being calculated properly for multisite registrations
