+++ app/presenters/reports/finance_data_report/initial_compliance_charge_row_presenter.rb
@@ -2,43 +2,17 @@
 
 module Reports
   module FinanceDataReport
-    class InitialComplianceChargeRowPresenter < BaseRegistrationRowPresenter
+    class InitialComplianceChargeRowPresenter < ComplianceChargeRowPresenter
       def charge_type
         "compliance_initial"
       end
 
-      def charge_amount
-        charge_amount = if @registration.multisite?
-                          @secondary_object.initial_compliance_charge_amount /
-                            @secondary_object.charge_detail.site_count
-                        else
-                          @secondary_object.initial_compliance_charge_amount
-                        end
-
-        display_pence_as_pounds_and_pence(pence: charge_amount,
-                                          hide_pence_if_zero: true)
-      end
-
-      def charge_band
-        @secondary_object.band.sequence
-      end
-
-      def exemption
-        @secondary_object.charge_detail.order.exemptions.select do |e|
-          e.band_id == @secondary_object.band_id && bucket_exemption_codes.exclude?(e.code)
-        end.map(&:code).sort.join(", ")
-      end
-
-      def balance
-        @total -= @secondary_object.initial_compliance_charge_amount
-        display_pence_as_pounds_and_pence(pence: @total,
-                                          hide_pence_if_zero: true)
-      end
-
-      private
-
-      def bucket_exemption_codes
-        @secondary_object.charge_detail.order.bucket&.exemptions&.map(&:code) || []
+      def charge_amount_in_pence
+        if @registration.multisite?
+          @secondary_object.initial_compliance_charge_amount / @secondary_object.charge_detail.site_count
+        else
+          @secondary_object.initial_compliance_charge_amount
+        end
       end
     end
   end
