+++ spec/presenters/reports/finance_data_report/additional_compliance_charge_row_presenter_spec.rb
@@ -24,10 +24,10 @@ module Reports
         end
       end
 
-      describe "#charge_amount" do
+      describe "#charge_amount_in_pence" do
         context "when registration is single site" do
-          it "returns the formatted charge amount" do
-            expect(presenter.charge_amount).to eq("20")
+          it "returns the formatted charge amount in pence" do
+            expect(presenter.charge_amount_in_pence).to eq(2000)
           end
         end
 
@@ -41,41 +41,8 @@ module Reports
           end
 
           it "returns the formatted charge amount divided by site count" do
-            expect(presenter.charge_amount).to eq("10")
+            expect(presenter.charge_amount_in_pence).to eq(1000)
           end
-        end
-      end
-
-      describe "#charge_band" do
-        it "returns the band sequence" do
-          expect(presenter.charge_band).to eq(2)
-        end
-      end
-
-      describe "#exemption" do
-        context "when there are no farmer exemptions" do
-          it "returns exemption code(s)" do
-            expect(presenter.exemption).to be_present
-          end
-        end
-
-        context "when there are farmer exemptions" do
-          let(:bucket) { build(:bucket, :farmer_exemptions) }
-          let(:order) { build(:order, :with_exemptions, :with_charge_detail, order_owner: account, bucket: bucket) }
-
-          it "does not include farmer exemption code(s)" do
-            farmer_exemption = order.exemptions.last
-            bucket.exemptions << farmer_exemption
-
-            expect(presenter.exemption).to be_present
-            expect(presenter.exemption).not_to include(farmer_exemption.code)
-          end
-        end
-      end
-
-      describe "#balance" do
-        it "returns the formatted balance amount" do
-          expect(presenter.balance).to eq("-31.50")
         end
       end
     end
